Move root router procedures
